arl-finger-add-pro/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
loguru==0.7.2
PyYAML==6.0.1
pyyml==0.0.2
requests==2.31.0
urllib3==2.1.0
